requirements.txt
# Core
openai>=1.12.0
python-dotenv>=1.0.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualisation
plotly>=5.18.0

# UI
gradio>=4.40.0

# Document processing
PyMuPDF>=1.23.0

# Testing
pytest>=7.4.0
